[WIP][MAINT] Add tests for painting commands in Graphics module (bis)

+++ src/DO/Graphics/DerivedQObjects/PaintingWindow.cpp
@@ -126,7 +126,7 @@ namespace DO {
     update();
   }
 
-  void PaintingWindow::drawCircle(const QPointF& center, float r,
+  void PaintingWindow::drawCircle(const QPointF& center, qreal r,
                                   const QColor& c, int penWidth)
   {
     painter_.setPen(QPen(c, penWidth));
